feat/파일분리, utill함수 적용
